Changes to enhance operation of mbzoltan, in particular to allow serial execution.  Tested for several of the partitioning methods allowed by zoltan.  Still needs to be run by mpiexec, though.
Also, couple of includes inserted that were necessary after creation
of MBEntityType.h, etc, and elimination of several warnings about
comparing to unsigned and unused variables.

MBCore: keep track of whether the constructor initialized MPI, and if
it did, then finalize in destructor.

MBInternals: make PROC_FROM_HANDLE return an unsigned int.

MBParallelComm: add assign_global_ids function, which assigns
contiguous ids over each topological dimension over mesh owned on all
processors (involves parallel communication step).  Global ids assumed
to start at 1.

MBZoltan: improved to work on single processor, taking number of
partitions as an argument.  In this case, no parallel communication is
done, though it still needs to be run with mpiexec (I think Zoltan
requires that).


git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@915 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ MBInternals.hpp
@@ -97,7 +97,7 @@ inline MBEntityType TYPE_FROM_HANDLE(MBEntityHandle handle)
 
 //! Get the type out of the handle.  Can do a simple shift because
 //! handles are unsigned (therefore shifting fills with zero's)
-inline int PROC_FROM_HANDLE(MBEntityHandle handle) 
+inline unsigned long PROC_FROM_HANDLE(MBEntityHandle handle) 
 {
   return (handle & MB_PROC_MASK) >> MB_ID_WIDTH;
 }
